@@ -4,8 +4,8 @@
   "private": true,
   "dependencies": {
     "axios": "^0.18.0",
-    "react": "^16.6.1",
-    "react-dom": "^16.6.1",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "react-redux": "^5.1.0",
     "react-router": "^4.3.1",
     "react-router-config": "^4.4.0-beta.6",
